--- source/mgnp/src/channel.rs
@@ -1,1 +1,0 @@
-
