+++ source/mgnp/src/client.rs
